Soft Robotic Hand for Hand Rehabilitation
This project aims to create a smart rehabilitation device that can be utilized at home and without the need to leave your house.
If the doctor following the patient’s condition needs to follow up on the patient ai a whole report will be sent to the doctor directly. 
To reduce the time it takes it takes to recover from approximately 1 year to 8 months.
Restoring Range of Motion: One of the primary objectives of hand rehabilitation is to restore the normal range of motion in the hand, wrist, and fingers.